Renamed "lib" to the more Pythonic "phyloref".
