+++ add-labels/add-labels.py
@@ -16,10 +16,10 @@ import json
 import os.path
 import sys
 
-# Add './lib' to lookup path.
-sys.path.append(os.path.join(os.path.dirname(__file__), "lib"))
+# Add './phyloref' to lookup path.
+sys.path.append(os.path.join(os.path.dirname(__file__), "phyloref"))
 
-from lib import PhyloreferenceTestSuite
+from phyloref import PhyloreferenceTestSuite
 
 __version__ = "0.1"
 __author__ = "Gaurav Vaidya"
